Location sharing using OpenStreetMap via osmdroid.
Use playProdOsmDebug variant for example to build it.

This implementation does not intend to remove all uses of Google APIs
from Signal. It is intended to add a drop-in replacement to the features
of Google Maps library that Signal uses, that replaces the "backend" of
all classes and methods to OpenStreetMap via osmdroid library. It does
not edit any call site code (except app/build.gradle to add a dimension
and flavors).